tests template added

+++ src/utils/dir-utils.ts
@@ -1,6 +1,8 @@
 import { existsSync, mkdirSync, writeFileSync } from 'fs';
 import * as vscode from 'vscode';
 import { resolve, normalize } from 'path';
+import { VsMessage } from './vsMessageException';
+import { NamingGenerator } from '../component-types/namingGenerator';
 export class DirUtils {
     private dir: string
     constructor(dir: string) {
@@ -15,29 +17,20 @@ export class DirUtils {
         return false
     }
 
-    sanitizeInput(): boolean {
-        const dir = this.dir;
-        let dirSplitted = dir.split("/");
-        let lastItem = dirSplitted.slice(dirSplitted.length - 1)[0];
-        console.log(dir, lastItem);
-        /*  if(lastItem.indexOf(".tsx") !== -1 || lastItem.indexOf(".jsx") !== -1) {
-             vscode.window.showInformationMessage("Directory name not valid");
-             return false
-         }  */
-        return true
-    }
-
     generateDirRecursive(dir: string) {
         var path = normalize(dir).replace(/\\/g, '/').split('/');
         for (var i = 1; i <= path.length; i++) {
             var segment = path.slice(0, i).join('/');
-            !existsSync(segment) ? mkdirSync(segment) : null ;
+            try {
+                existsSync(segment) ?  null : mkdirSync(segment) ;
+            } catch(e) {
+                console.log("ERROR", e);
+            }
+            
         }
     }
-    generateDir() {
-        
+    generateDir() {   
         const dirFolder = resolve(this.dir, "../");
-        console.log("DIR",this.dir, dirFolder);
         this.generateDirRecursive(dirFolder);
     }
 
@@ -48,7 +41,25 @@ export class DirUtils {
             
         }
         this.generateDir();
-        writeFileSync(dir, content, { encoding: 'utf8' });
+        if(!existsSync(dir)) {
+            writeFileSync(dir, content, { encoding: 'utf8' });
+        }else {
+            throw new VsMessage("This file already exists", true);
+        }
+        
+        return true
+    }
+
+    generateTestFile(content: string, naming: NamingGenerator): boolean {
+        const dir = resolve(this.dir, "../") + "/" + naming.name + ".spec." + naming.extension;
+        console.log("DIRT", dir);
+
+        if(!existsSync(dir)) {
+            writeFileSync(dir, content, { encoding: 'utf8' });
+        }else {
+            throw new VsMessage("This file already exists", true);
+        }
+        
         return true
     }
 }
